Handle images done

+++ app/Traits/FileUploadTrait.php
@@ -7,7 +7,7 @@ use Illuminate\Http\Request;
 
 trait FileUploadTrait
 {
-    function uploadImage(Request $request, string $inputName, string $oldPath = null, string $path = '/uploads'): ?string
+    function uploadImage(Request $request, string $inputName, ?string $oldPath = null, string $path = '/uploads'): ?string
     {
         //Hàm này kiểm tra xem ảnh có được upload lên không
         if ($request->hasFile($inputName)) {
@@ -21,8 +21,10 @@ trait FileUploadTrait
             //File sẽ được di chuyển tới thư mục uploads cụ thể hơn /public/uploads/media_64fd0d674ec2f.jpg
             $image->move(public_path($path), $imageName);
 
+            $emptyFolder = "/default";
+
             //Xóa ảnh gốc từ storage
-            if ($oldPath && File::exists(public_path($oldPath))) {
+            if ($oldPath && File::exists(public_path($oldPath)) && strpos($oldPath, $emptyFolder) !== 0) {
                 File::delete(public_path($oldPath));
             }
             //Trả về đường dẫn sau khi thành công /uploads/media_64fd0d674ec2f.jpg
